% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_outliers.R
\name{boxplot_outliers}
\alias{boxplot_outliers}
\title{Boxplot outlier identification method}
\usage{
boxplot_outliers(data, depths_vector, inf_factor = 1.5)
}
\arguments{
\item{data}{The data input. If you are inputting a data given from sim_out, then remember to drop the last column}

\item{depths_vector}{Vector of depth for a given data set}

\item{inf_factor.}{The parameter by which the interquartile rank should be inflated (default is 1.5)}
}
\value{
functions for upper box and whiskers, as well as the ids of the outliers
false positive, false negative and true negative counts for each iteration of the simulation.
}
\description{
Function that, for a given input dataset and a vector of depths, outputs the outliers as
observation number for the dataset
}
