% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_matrix.R
\name{c_matrix}
\alias{c_matrix}
\title{Confusion matrix for the outlier detection algorithm in simulation}
\usage{
c_matrix(data, method)
}
\arguments{
\item{data}{The data input. If you are inputting a data given from sim_out, then remember to drop the last column}

\item{class}{A string vector containing the types of outliers.}

\item{moderate}{(LOGICAL) Is it the moderate or the extreme scenario?}

\item{perc_out}{a vector of two entries  c(amplitude, shape) for the proportion of
observations outlying in each case.}
}
\value{
A named matrix which columns are, respectively, the true positive,
false positive, false negative and true negative counts for each iteration of the simulation.
}
\description{
This is an intermediate function to present the results of the simulation exercise
for the heterogeneous outlier detection in functional data.
}
