% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_matrix.R, R/compiler_functional.R
\name{compiler_out}
\alias{compiler_out}
\title{Confusion matrix indicators for outlier detection in functional data}
\usage{
compiler_out(data)

compiler_out(data)
}
\arguments{
\item{data}{a Nxt+1 dataframe with N observations of T discrete units, and one column of flags in the t+1 position.}

\item{keep}{logical / if TRUE, return details of each iteration in a dataframe.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A dataframe with a first column of the true flags, and further columns with the flags from the methods that are studied
}\if{html}{\out{</div>}}
}
\description{
This is a function to implement and summarize the results of outlier detection in one sample of functional data, contaminated with
outliers of different types.

This function uses \code{funData} package to generate univariate functional observations
with different types of outliers and different degrees of contamination.
}
