% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfd_sim.R, R/mfd_sim1.R
\name{mfd_sim}
\alias{mfd_sim}
\title{Simulation algorithm for multivariate functional data based on Ieva and Paganoni (2017)}
\usage{
mfd_sim(
  N,
  mu,
  grid = FALSE,
  covar = "sq",
  rho = 0,
  method = c("svd", "chol", "eigen"),
  delta = 0.5
)

mfd_sim(
  N,
  mu,
  grid = FALSE,
  covar = "sq",
  rho = 0,
  method = c("svd", "chol", "eigen"),
  delta = 0.5
)
}
\arguments{
\item{N}{Number of functional observations}

\item{mu}{Matrix with L rows and P columns containing the mean curves for the L components at discrete grid points}

\item{grid}{If you want to specify your own grid, the vector of the grid. Length must coincide with the number of columns in mu. Default is FALSE, in which case a \link{0,1} grid with even spacing
over P points is implemented.}

\item{covar}{List of L covariance matrices, one for each variable, each with P x P dimensions}

\item{rho}{Vector of cross correlations among L components}

\item{method}{Simulation method}
}
\value{
A list of simulated functional data

A list of simulated functional data
}
\description{
This function simulates multivariate functional data using Eigen, Cholesky and Singular
Value Decomposition, based on the method proposed by Ieva and Paganoni (2017) and implemented
initially in the library 'roahd', but with some changes enabling data exploration.

This function simulates multivariate functional data using Eigen, Cholesky and Singular
Value Decomposition, based on the method proposed by Ieva and Paganoni (2017) and implemented
initially in the library 'roahd', but with some changes enabling data exploration.
}
