% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfd_sim_c.R
\name{mfd_sim_c}
\alias{mfd_sim_c}
\title{Simulation algorithm for multivariate functional data from Claeskens et al 2014}
\usage{
mfd_sim_c(
  N,
  grid1,
  grid2,
  der = FALSE,
  covar = "sq",
  method = c("svd", "chol", "eigen")
)
}
\arguments{
\item{N}{Number of functional observations}

\item{der}{Boolean indicating if numerical derivatives should be calculating}

\item{covar}{Covariance method. Could be absolute exponential decay ("abs") or cuadratic exponential decay ("sq")}

\item{method}{Simulation method}

\item{grid, grid2}{Grids of discrete points in which the function is mapped (1) and the heteroskedasticity component (2)}

\item{delta}{Delta parameter of the covariance kernel}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A list of simulated functional data
}\if{html}{\out{</div>}}
}
\description{
This function simulates multivariate functional data using either a mixture of function without
covariance specification, or a heteroskedastic function from a mean and random noise
}
