% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfd_sim_h.R
\name{mfd_sim_h}
\alias{mfd_sim_h}
\title{Simulation algorithm for multivariate functional data from Claeskens et al 2014}
\usage{
mfd_sim_h(
  N,
  P,
  grid,
  mu,
  der = FALSE,
  covar = "sq",
  method = c("svd", "chol", "eigen"),
  het = TRUE,
  lognormal = TRUE,
  delta = 0.25
)
}
\arguments{
\item{N}{Number of functional observations}

\item{P}{Grid for the final data}

\item{der}{Boolean indicating if numerical derivatives should be calculated}

\item{covar}{Covariance method. Could be absolute exponential decay ("abs") or cuadratic exponential decay ("sq")}

\item{method}{Simulation method}

\item{het}{Heteroskedasticity. Default is TRUE. If False, D is an identity matrix}

\item{lognormal}{if TRUE, the distribution is assumed to be lognormal. If FALSE, assumed to be normal.}

\item{delta}{Delta parameter of the covariance kernel}

\item{grid1, grid2}{Numeric (Seq) grids of discrete points in which the function is mapped (1) and the heteroskedasticity component (2)}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{   A list of simulated functional data
}\if{html}{\out{</div>}}
}
\description{
This function simulates multivariate functional data based on a mean function and a covariance matrix
}
