% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msplot_simple.R
\name{msplot_simple}
\alias{msplot_simple}
\title{Simplification of msplot}
\usage{
msplot_simple(data)
}
\arguments{
\item{data}{The data input. If you are inputting a data given from sim_out, then remember to drop the last column}
}
\value{
data and outliers
false positive, false negative and true negative counts for each iteration of the simulation.
}
\description{
Function to simplify the msplot
}
