% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_out.R
\name{sim_out}
\alias{sim_out}
\title{Simulation algorithm for multivariate functional data for analyzing heterogeneous outliers}
\usage{
sim_out(nrep, N, p, moderate = TRUE, perc_out = c(0, 0))
}
\arguments{
\item{nrep}{Number of iterations to be performed}

\item{N}{Number of observations at each iteration}

\item{p}{Number of discretizations for t}

\item{moderate}{(LOGICAL) Is it the moderate or the extreme scenario?}

\item{perc_out}{a vector of two entries  c(amplitude, shape) for the proportion of
observations outlying in each case.}
}
\value{
A list of dimension nrep of simulated functional
}
\description{
This function uses funData package to generate univariate functional observations
with different types of outliers and different degrees of contamination
}
