% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFunDataOut.R
\name{simFunDataOut}
\alias{simFunDataOut}
\title{Simulation algorithm for univariate functional data with heterogeneous outliers}
\usage{
simFunDataOut(
  N,
  Nt,
  mean,
  prop,
  M = 10,
  eFunType = "Fourier",
  eValType = "exponential"
)
}
\arguments{
\item{N}{Number of functional observations.}

\item{Nt}{Number of discrete time points.}

\item{mean}{A list of mean functions for base data and outlier types.}

\item{prop}{A vector of proportions for base data and outlier types.}

\item{M}{(Numeric) the number of eignevalues / eigenfunctions to be considered in simFunData}

\item{eFunType}{(Character) the type of eigenfunction to be simulated for the random term. Types include "Poly", "PolyHigh", "Fourier", "FourierLin", and "Wiener"}

\item{eValType}{(Character) The type of decay for the eigenvalues in the random component. Could be "linear", "exponential" or "wiener"}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ An Nx(P+1) matrix containing functional data and a class membership indicator for base data and outlier types.
}\if{html}{\out{</div>}}
}
\description{
This function uses \code{funData} package to generate univariate functional observations
with different types of outliers and different degrees of contamination.
}
