% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_results.R
\name{sum_results}
\alias{sum_results}
\title{Summary of simulation results}
\usage{
sum_results(results)
}
\arguments{
\item{results}{matrix given from the c_matrix function. Its columns are the counting of true and
false positive, false and true negatives.}
}
\value{
a vector of mean and standard deviation ,
false positive, false negative and true negative counts for each iteration of the simulation.
}
\description{
Basic summary of the confusion matrix.
}
