% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufd_sim.R
\name{ufd_sim}
\alias{ufd_sim}
\title{Simulation algorithm for multivariate functional data based on Ieva and Paganoni (2017)}
\usage{
ufd_sim(
  N,
  mu,
  grid = FALSE,
  covar = "sq",
  method = c("chol", "eigen"),
  delta = 0.5
)
}
\arguments{
\item{N}{Number of functional observations}

\item{mu}{Matrix with 1 row and P columns containing the mean curve}

\item{grid}{If you want to specify your own grid, the vector of the grid. Length must coincide with the number of elements. Default is FALSE, in which case a \link{0,1} grid with even spacing
over P points is implemented.}

\item{covar}{Either a covariance matrix or a type of exponential decay, either "sq" for square or "abs" for absolute exponential decay}

\item{method}{Factorization method for the covariance matrix. Either "eigen" for eigendecomposition or "chol" for Choleski factorization.}
}
\value{
a list with "x", the simulated data, "covar", the covariance matrix that was used as input, "factorized covar" the covariance matrix after factorization, "Z_matrix", each of the elements of factorization.
}
\description{
This function simulates multivariate functional data using Eigen, Cholesky and Singular
Value Decomposition, based on the method proposed by Ieva and Paganoni (2017) and implemented
initially in the library 'roahd', but with some changes enabling data exploration.
}
