% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_depth.R
\name{uni_depth}
\alias{uni_depth}
\title{Basic functions for univariate depth calculation}
\usage{
uni_depth(data, depth = "maha")
}
\arguments{
\item{data}{An NxP Matrix, from which a column-wise depth is calculated}

\item{depth}{What depth is going to be used. Supported are Mahalanobis, Projection, Univariate Tukey and Simplicial}
}
\description{
This function calculates column-wise univariate depths that would serve as inputs for integrate and geometric
multivariate and functional depths.
}
