% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler_functional.R
\name{c_matrix}
\alias{c_matrix}
\title{Simulation algorithm for univariate functional data with heterogeneous outliers}
\usage{
c_matrix(data)
}
\arguments{
\item{data}{a Nxt+1 dataframe with N observations of T discrete units, and one column of flags in the t+1 position.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{ A dataframe with a first column of the true flags, and further columns with the flags from the methods that are studied
}\if{html}{\out{</div>}}
}
\description{
This function uses \code{funData} package to generate univariate functional observations
with different types of outliers and different degrees of contamination.
}
