% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_matrix.R
\name{compiler_out}
\alias{compiler_out}
\title{Confusion matrix indicators for outlier detection in functional data}
\usage{
compiler_out(data, keep)
}
\arguments{
\item{data}{data Output of simFunDataOut}

\item{keep}{logical / if TRUE, return details of each iteration in a dataframe.}
}
\description{
This is a function to implement and summarize the results of outlier detection in one sample of functional data, contaminated with
outliers of different types.
}
