% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfd_sim.R
\name{mfd_sim}
\alias{mfd_sim}
\title{Simulation algorithm for multivariate functional data from Ieva and Paganoni (2017)}
\usage{
mfd_sim(N, mu, covar = NULL, rho = 0, method = c("svd", "chol", "eigen"))
}
\arguments{
\item{N}{Number of functional observations}

\item{mu}{Matrix with L rows and P columns containing the mean curves for the L components at discrete grid points}

\item{covar}{List of L covariance matrices, one for each variable, each with P x P dimensions}

\item{rho}{Vector of cross correlations among L components}

\item{method}{Simulation method}
}
\value{
A list of simulated functional data
}
\description{
This function simulates multivariate functional data using Eigen, Cholesky and Singular
Value Decomposition, based on the method proposed by Ieva and Paganoni (2017) and implemented
In the library 'roahd
}
